requirements.txt
celery
redis
requests
beautifulsoup4
sqlalchemy
psycopg2-binary
python-dotenv
lxml
